WIP: Simplify `volume_rendering_xr` Dockerfile
Simplifies the `volume_rendering_xr` Dockerfile for ease of development.

Under previous behavior the `volume_rendering_xr` container build
process was split into multiple components for the purpose of separating
development and release containers with a common base. However, the
release container is not currently available on HoloHub.

This PR consolidates the development container build process for
compatibility with the HoloHub `build_and_run` infrastructure, pending
additional alignment on the release container.

WIP:
- Windrunner is currently disabled pending a new build release.
- `build_and_run` is untested.